Add comp folder, move files, add dependencies
