+++ components/SponsorSection.module.css
@@ -1,5 +1,3 @@
-@import "../../static/stylesheets/_variables.scss";
-
 .sponsorTitle {
   padding-bottom: 4em;
 }
