+++ package.json
@@ -10,6 +10,8 @@
   "dependencies": {
     "next": "10.1.3",
     "react": "17.0.2",
-    "react-dom": "17.0.2"
+    "react-bootstrap": "^1.6.0",
+    "react-dom": "17.0.2",
+    "react-scroll": "^1.8.2"
   }
 }
